refactor: move stream config to core
