+++ bililive-core/src/lib.rs
@@ -1,4 +1,5 @@
 #![allow(clippy::cast_lossless, clippy::cast_possible_truncation)]
 
-pub mod packet;
+pub mod config;
 pub mod errors;
+pub mod packet;
