+++ bililive/src/config.rs
@@ -6,48 +6,6 @@ use rand::distributions::Uniform;
 use rand::{thread_rng, Rng};
 use stream_reconnect::ReconnectOptions;
 
-/// The configuration for bilibili live stream connection.
-#[derive(Debug, Clone)]
-pub struct StreamConfig(Box<Inner>);
-
-impl StreamConfig {
-    pub fn new(room_id: u64, uid: u64, token: String, servers: Vec<String>) -> Self {
-        Self(Box::new(Inner {
-            room_id,
-            uid,
-            token,
-            servers,
-        }))
-    }
-}
-
-impl StreamConfig {
-    pub const fn room_id(&self) -> u64 {
-        self.0.room_id
-    }
-    pub const fn uid(&self) -> u64 {
-        self.0.uid
-    }
-    pub fn token(&self) -> &str {
-        &self.0.token
-    }
-    pub fn servers(&self) -> &[String] {
-        &self.0.servers
-    }
-}
-
-#[derive(Debug, Clone)]
-struct Inner {
-    /// Live room id (long version).
-    room_id: u64,
-    /// Live room user id.
-    uid: u64,
-    /// Danmaku server token.
-    token: String,
-    /// Danmaku server urls.
-    servers: Vec<String>,
-}
-
 /// The configuration for retry behavior.
 #[derive(Clone)]
 pub struct RetryConfig(ReconnectOptions);
